updated the function to use size_t

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -12,9 +12,9 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
 		action(array[i]);
 	}
